app/requirements.txt
accelerate==1.1.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiosignal==1.3.1
altair==5.4.0
annotated-types==0.7.0
anyio==4.6.2.post1
asgiref==3.7.2
attrs==24.2.0
Babel==2.12.1
beautifulsoup4==4.12.2
blinker==1.8.2
cachetools==5.5.0
certifi==2023.7.22
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
colorama==0.4.6
contourpy==1.2.0
crispy-bootstrap5==0.7
cryptography==42.0.8
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
Django==4.2.2
django-bootstrap-icons==0.8.3
django-bootstrap-v5==1.0.11
django-bootstrap5==23.3
django-cleanup==8.0.0
django-crispy-forms==2.0
django-jazzmin==2.6.0
django-phone-field==1.8.1
django-phonenumber-field==7.1.0
django-resized==1.0.2
et_xmlfile==2.0.0
fastapi==0.115.4
filelock==3.16.1
fonttools==4.47.2
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
huggingface-hub==0.26.2
idna==3.4
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.8.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.5.2
networkx==3.3
numpy==1.26.4
openai==0.28.0
opencv-python==4.10.0.84
openpyxl==3.1.5
packaging==23.2
pandas==2.2.3
pdfminer.six==20231228
pdfplumber==0.11.2
phonenumbers==8.13.15
Pillow==10.0.0
propcache==0.2.0
protobuf==5.27.3
psutil==6.1.0
pyarrow==17.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydeck==0.9.1
Pygments==2.18.0
PyMuPDF==1.24.7
PyMuPDFb==1.24.6
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.30.0
python-dateutil==2.8.2
python-docx==1.1.2
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
reportlab==4.2.2
requests==2.32.3
rich==13.7.1
rpds-py==0.20.0
safetensors==0.4.5
scikit-learn==1.5.2
scipy==1.12.0
sentence-transformers==3.2.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.4.1
sqlparse==0.4.4
starlette==0.41.2
streamlit==1.37.1
sympy==1.13.1
tenacity==8.5.0
threadpoolctl==3.5.0
tokenizers==0.20.1
toml==0.10.2
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1
tornado==6.4.1
tqdm==4.66.6
transformers==4.46.1
typing_extensions==4.12.2
tzdata==2023.3
urllib3==2.0.4
uvicorn==0.32.0
watchdog==4.0.2
whitenoise==6.5.0
xxhash==3.5.0
yarl==1.17.1
